Move DatabaseOptions to separate file
